Алгоритм обратного распространения
Доавблен аглоритм обратного распространения для сети и слоёв. Исправлены ошибки.

+++ comm_funcs.h
@@ -4,17 +4,17 @@
 #include <math.h>
 #include <random>
 
-float gaussrand_t(float MO, float sko);
-void gen_array(const float N, const float M, const int Z, float *array);
+ double gaussrand_t( double MO,  double sko);
+void gen_array(const  double N, const  double M, const int Z,  double *array);
 int8_t Parser(std::string& input,const std::string* possible_values_list, uint8_t num_of_possible);
 std::vector<std::string>& Get_words(std::string& str, std::vector<std::string>& arr,std::string delim);
 
 //Функции активации
-float SIGMOID(float S);
-float ReLU(float S);
-float Linear(float S);
-float SoftMax(float S);
+ double SIGMOID( double S);
+ double ReLU( double S);
+ double Linear( double S);
+ double SoftMax( double S);
 //Производные функций активации
-float DSIGMIOD(float S);
+ double DSIGMIOD( double S);
 
 #endif // RANDOM_NUM_H
